moved chat for mobile rwd out of Main.jsx
