+++ web/src/home/Home.css
@@ -69,12 +69,35 @@ body {
     z-index: 10;
 }
 
+@media only screen and (min-width: 1100px) {
+    .sponsorsContainerHome {
+        display: none;
+    }
+}
+
 @media only screen and (max-width: 1100px) {
     .menu {
         display: none;
     }
 
+    .divContainer {
+        display: block;
+        height: 100vh;
+        overflow-x: auto;
+    }
+
+    .sponsorsContainer {
+        width: 100vw;
+    }
+
     .chatContainerHome {
-        display: none;
+        overflow: auto;
+        width: 100%;
+        height: 50rem;
+        z-index: 10;
+    }
+
+    .subpageContainer {
+        width: 100vw;
     }
 }
