+++ web/src/pages/main/Main.css
@@ -52,7 +52,7 @@
     .timerContainer {
         text-align: center;
         margin-top: 60px;
-        margin-bottom: 0px;
+        margin-bottom: 0;
         z-index: 10;
     }
 
@@ -65,10 +65,7 @@
         z-index: 10;
     }
 
-    .chatContainer {
-        overflow: auto;
-        width: 100%;
-        height: 75%;
-        z-index: 10;
+    .sponsorsContainer {
+        display: none;
     }
 }
